requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
Markdown==3.5.1
PyYAML==6.0.1
requests==2.31.0
